More complex networking
Refactors a lot of networking stuff. Two big things to keep in mind. This (theoretically) introduces the possibility that we close an image transaction buffer before all the image packets come through but that didn't seem to be a problem for android's implementation of TCP. This is also the best we'll get because of how TCP works as an abstraction. Also, I added a bit of code to reload the UI after getting an image which breaks when you send a message to 127.0.0.1. Oh well.