Prelude to Post persistance
Everything we need to turn on feed persistence is here. But it seems that our plan of storing images as blobs requires a lot more twiddling. There's some issue of SQL cursors being too small to actually read the images which causes really annoying UI errors so its turned off.